use batch send technique in fifo market
