+++ .vscode/settings.json
@@ -51,7 +51,7 @@
   /** END FILE WATCHER OPTIONS */
 
   /** BEGIN SOLIDITY OPTIONS */
-  "solidity.compileUsingRemoteVersion": "v0.8.12+commit.f00d7308",
+  "solidity.compileUsingRemoteVersion": "0.8.13",
   /** END SOLIDITY OPTIONS */
 
   /** BEGIN CYPRESS OPTIONS */
