--- deployments/goerli/.chainId
@@ -1,1 +1,0 @@
-5
