--- deployments/mainnet/.chainId
@@ -1,1 +1,0 @@
-1
