--- deployments/mainnet/NORI.json
@@ -1,5 +1,0 @@
-{
-  "address": "0x961760Ad1bEd52bf4d79aa4b1558E7F9d72071e4",
-  "numDeployments": 1,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
