--- deployments/mumbai/BridgedPolygonNORI.json
@@ -1,5 +1,0 @@
-{
-  "address": "0x6E570A15C0C39b5F29388DEEa6ebda8CDDc40587",
-  "numDeployments": 2,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
