--- deployments/mumbai/Certificate.json
@@ -1,5 +1,0 @@
-{
-  "address": "0xF5bafc8B26Fb38ea788Bc6eAf3d84ba3bc4Ca91C",
-  "numDeployments": 2,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
