--- deployments/mumbai/FIFOMarket.json
@@ -1,5 +1,0 @@
-{
-  "address": "0x61819E8238A3493547Ae1131dc657D87F62833FD",
-  "numDeployments": 1,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
