--- deployments/mumbai/LockedNORI.json
@@ -1,5 +1,0 @@
-{
-  "address": "0xCdcB43CB7B668F0C1cA04FE4b60DA7F8C62Be393",
-  "numDeployments": 1,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
