--- deployments/mumbai/Removal.json
@@ -1,5 +1,0 @@
-{
-  "address": "0xa484c59eB391afFd88C7de0dBF4B6E7e2A9E84fE",
-  "numDeployments": 1,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
