--- deployments/polygon/BridgedPolygonNORI.json
@@ -1,5 +1,0 @@
-{
-  "address": "0x8cf6E82919f69aE382DEf8f94e581a43Ce1E70C1",
-  "numDeployments": 1,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
