--- deployments/polygon/LockedNORI.json
@@ -1,5 +1,0 @@
-{
-  "address": "0xCcFfFA6c2a030821331cC113b63babDC60BfF82A",
-  "numDeployments": 1,
-  "metadata": "{\"compiler\":{\"version\":\"0.8.13\"}"
-}
